Add v3, solución backend documentos

+++ app/Http/Controllers/FuentesSecundarias/DocumentoAutoevaluacionController.php
@@ -127,25 +127,26 @@ class DocumentoAutoevaluacionController extends Controller
      */
     public function store(DocumentosAutoevaluacionRequest $request)
     {
+        $proceso = Proceso::findOrFail(session()->get('id_proceso'));
 
+        if ($proceso->FK_PCS_Fase == 4) {
+            if ($request->hasFile('archivo')) {
+                $archivo = $request->file('archivo');
+                $nombre = pathinfo($archivo->getClientOriginalName(), PATHINFO_FILENAME);
+                $extension = $archivo->getClientOriginalExtension();
+                $url = Storage::url($archivo->store('public/documentos_autoevaluacion'));
 
-        if ($request->hasFile('archivo')) {
-            $archivo = $request->file('archivo');
-            $nombre = pathinfo($archivo->getClientOriginalName(), PATHINFO_FILENAME);
-            $extension = $archivo->getClientOriginalExtension();
-            $url = Storage::url($archivo->store('public/documentos_autoevaluacion'));
+                $archivos = new Archivo();
+                $archivos->ACV_Nombre = $nombre;
+                $archivos->ACV_Extension = $extension;
+                $archivos->ruta = $url;
+                $archivos->save();
 
-            $archivos = new Archivo();
-            $archivos->ACV_Nombre = $nombre;
-            $archivos->ACV_Extension = $extension;
-            $archivos->ruta = $url;
-            $archivos->save();
+                $id_archivo = $archivos->PK_ACV_Id;
+            }
 
-            $id_archivo = $archivos->PK_ACV_Id;
-        }
-
-        $documento_auto = new DocumentoAutoevaluacion();
-        $documento_auto->fill($request->only(['IDO_Nombre',
+            $documento_auto = new DocumentoAutoevaluacion();
+            $documento_auto->fill($request->only(['IDO_Nombre',
             'DOA_Numero',
             'DOA_Anio',
             'DOA_Link',
@@ -154,17 +155,23 @@ class DocumentoAutoevaluacionController extends Controller
             'DOA_ContenidoAdicional',
             'DOA_Observaciones']));
 
-        $documento_auto->FK_DOA_Archivo = isset($id_archivo) ? $id_archivo : null;
-        $documento_auto->FK_DOA_IndicadorDocumental = $request->get('PK_IDO_Id');
-        $documento_auto->FK_DOA_TipoDocumento = $request->get('PK_TDO_Id');
-        $documento_auto->FK_DOA_Dependencia = $request->get('PK_DPC_Id');
-        $documento_auto->FK_DOA_Proceso = session()->get('id_proceso');
-        $documento_auto->save();
+            $documento_auto->FK_DOA_Archivo = isset($id_archivo) ? $id_archivo : null;
+            $documento_auto->FK_DOA_IndicadorDocumental = $request->get('PK_IDO_Id');
+            $documento_auto->FK_DOA_TipoDocumento = $request->get('PK_TDO_Id');
+            $documento_auto->FK_DOA_Dependencia = $request->get('PK_DPC_Id');
+            $documento_auto->FK_DOA_Proceso = session()->get('id_proceso');
+            $documento_auto->save();
 
-        return response(['msg' => 'El documento se ha registrado correctamente.',
+            return response(['msg' => 'El documento se ha registrado correctamente.',
             'title' => '¡Registro exitoso!'
         ], 200)// 200 Status Code: Standard response for successful HTTP request
         ->header('Content-Type', 'application/json');
+        }
+         return response([
+                'errors' => ['El proceso se debe encontrar en fase de captura de datos para poder subir documentos.'],
+                'title' => '¡Error!'
+            ], 422)// 200 Status Code: Standard response for successful HTTP request
+            ->header('Content-Type', 'application/json');
     }
 
     /**
@@ -224,44 +231,42 @@ class DocumentoAutoevaluacionController extends Controller
 
         $documento = DocumentoAutoevaluacion::findOrFail($id);
         $this->authorize('autorizar', $documento->FK_DOA_Proceso);
+        $proceso = Proceso::find($documento->FK_DOA_Proceso);
 
-        if ($request->hasFile('archivo')) {
-            $archivo = $request->file('archivo');
-            $nombre = pathinfo($archivo->getClientOriginalName(), PATHINFO_FILENAME);
-            $extension = $archivo->getClientOriginalExtension();
-            $url = Storage::url($archivo->store('public/documentos_autoevaluacion'));
+         if ($proceso->FK_PCS_Fase == 4 || $proceso->FK_PCS_Fase == 5) {
+             if ($request->hasFile('archivo')) {
+                 $archivo = $request->file('archivo');
+                 $nombre = pathinfo($archivo->getClientOriginalName(), PATHINFO_FILENAME);
+                 $extension = $archivo->getClientOriginalExtension();
+                 $url = Storage::url($archivo->store('public/documentos_autoevaluacion'));
 
-            if ($documento->archivo) {
-                $ruta = str_replace('storage', 'public', $documento->archivo->ruta);
-                Storage::delete($ruta);
-                $archivos = Archivo::findOrfail($documento->FK_DOA_Archivo);
-                $archivos->ACV_Nombre = $nombre;
-                $archivos->ACV_Extension = $extension;
-                $archivos->ruta = $url;
-                $archivos->update();
-                $id_archivo = $archivos->PK_ACV_Id;
+                 if ($documento->archivo) {
+                     $ruta = str_replace('storage', 'public', $documento->archivo->ruta);
+                     Storage::delete($ruta);
+                     $archivos = Archivo::findOrfail($documento->FK_DOA_Archivo);
+                     $archivos->ACV_Nombre = $nombre;
+                     $archivos->ACV_Extension = $extension;
+                     $archivos->ruta = $url;
+                     $archivos->update();
+                     $id_archivo = $archivos->PK_ACV_Id;
+                 } else {
+                     $archivos = new Archivo();
+                     $archivos->ACV_Nombre = $nombre;
+                     $archivos->ACV_Extension = $extension;
+                     $archivos->ruta = $url;
+                     $archivos->save();
 
+                     $id_archivo = $archivos->PK_ACV_Id;
+                 }
+             }
+             if ($request->get('DOA_Link') != null && $documento->archivo) {
+                 $documento->FK_DOA_Archivo = null;
+                 $borraArchivo = true;
+                 $ruta = $documento->archivo->ruta;
+                 $id = $documento->FK_DOA_Archivo;
+             }
 
-            } else {
-                $archivos = new Archivo();
-                $archivos->ACV_Nombre = $nombre;
-                $archivos->ACV_Extension = $extension;
-                $archivos->ruta = $url;
-                $archivos->save();
-
-                $id_archivo = $archivos->PK_ACV_Id;
-
-            }
-
-        }
-        if($request->get('DOA_Link') != null && $documento->archivo){
-            $documento->FK_DOA_Archivo = null;
-            $borraArchivo = true;
-            $ruta = $documento->archivo->ruta;
-            $id = $documento->FK_DOA_Archivo;
-        }
-
-        $documento->fill($request->only([
+             $documento->fill($request->only([
             'IDO_Nombre',
             'DOA_Numero',
             'DOA_Anio',
@@ -271,27 +276,34 @@ class DocumentoAutoevaluacionController extends Controller
             'DOA_ContenidoAdicional',
             'DOA_Observaciones'
         ]));
-        if (isset($id_archivo)){
-            $documento->FK_DOA_Archivo = $id_archivo;
-        }
+             if (isset($id_archivo)) {
+                 $documento->FK_DOA_Archivo = $id_archivo;
+             }
 
         
-        $documento->FK_DOA_IndicadorDocumental = $request->get('PK_IDO_Id');
-        $documento->FK_DOA_TipoDocumento = $request->get('PK_TDO_Id');
-        $documento->FK_DOA_Dependencia = $request->get('PK_DPC_Id');
-        $documento->update();
+             $documento->FK_DOA_IndicadorDocumental = $request->get('PK_IDO_Id');
+             $documento->FK_DOA_TipoDocumento = $request->get('PK_TDO_Id');
+             $documento->FK_DOA_Dependencia = $request->get('PK_DPC_Id');
+             $documento->update();
 
-        if ($borraArchivo) {
-            $ruta = str_replace('storage', 'public', $ruta);
-            Storage::delete($ruta);
-            Archivo::destroy($id);
-        }
+             if ($borraArchivo) {
+                 $ruta = str_replace('storage', 'public', $ruta);
+                 Storage::delete($ruta);
+                 Archivo::destroy($id);
+             }
 
 
-        return response(['msg' => 'El Indicador documental ha sido modificado exitosamente.',
+             return response(['msg' => 'El Indicador documental ha sido modificado exitosamente.',
             'title' => 'Indicador Modificado!'
         ], 200)// 200 Status Code: Standard response for successful HTTP request
         ->header('Content-Type', 'application/json');
+        }
+        return response([
+                'errors' => ['El proceso se debe encontrar en fase de captura de datos o de consolidación para poder modificar la información de los documentos.'],
+                'title' => '¡Error!'
+            ], 422)// 200 Status Code: Standard response for successful HTTP request
+            ->header('Content-Type', 'application/json');
+
     }
 
     /**
